added contact motion
